Refactor(Data): 데이터 레이어 관련 소스 코드 :data 모듈로 마이그레이션
