--- app/src/main/java/com/gyleedev/githubsearch/data/remote/AccessService.kt
@@ -1,19 +1,0 @@
-package com.gyleedev.githubsearch.data.remote
-
-import com.gyleedev.githubsearch.data.remote.response.GithubAccessResponse
-import retrofit2.Response
-import retrofit2.http.Field
-import retrofit2.http.FormUrlEncoded
-import retrofit2.http.Headers
-import retrofit2.http.POST
-
-interface AccessService {
-    @FormUrlEncoded
-    @POST("login/oauth/access_token")
-    @Headers("Accept: application/json")
-    suspend fun getAccessToken(
-        @Field("client_id") clientId: String,
-        @Field("client_secret") clientSecret: String,
-        @Field("code") code: String,
-    ): Response<GithubAccessResponse>
-}
