--- app/src/main/java/com/gyleedev/githubsearch/data/remote/Qualifiers.kt
@@ -1,15 +1,0 @@
-package com.gyleedev.githubsearch.data.remote
-
-import javax.inject.Qualifier
-
-@Qualifier
-@Retention(AnnotationRetention.BINARY)
-annotation class TypeAccess
-
-@Qualifier
-@Retention(AnnotationRetention.BINARY)
-annotation class TypeApi
-
-@Qualifier
-@Retention(AnnotationRetention.BINARY)
-annotation class TypeRevoke
