--- app/src/main/java/com/gyleedev/githubsearch/data/remote/response/GithubAccessResponse.kt
@@ -1,5 +1,0 @@
-package com.gyleedev.githubsearch.data.remote.response
-
-import com.google.gson.annotations.SerializedName
-
-data class GithubAccessResponse(@SerializedName("access_token") val accessToken: String)
