--- app/src/main/java/com/gyleedev/githubsearch/data/remote/response/RevokeResponse.kt
@@ -1,5 +1,0 @@
-package com.gyleedev.githubsearch.data.remote.response
-
-import com.google.gson.annotations.SerializedName
-
-data class RevokeResponse(@SerializedName("Status") val status: Int)
