+++ data/src/main/java/com/gyleedev/data/PreferenceUtil.kt
@@ -1,0 +1,25 @@
+package com.gyleedev.data
+
+import android.content.Context
+import android.content.SharedPreferences
+import dagger.hilt.android.qualifiers.ApplicationContext
+
+class PreferenceUtil(@ApplicationContext context: Context) {
+
+    private val accessPreference: SharedPreferences =
+        context.getSharedPreferences("AccessToken", Context.MODE_PRIVATE)
+
+    fun getString(key: String = "Token", defValue: String): String = accessPreference.getString(key, defValue) as String
+
+    fun setString(key: String = "Token", str: String) {
+        accessPreference.edit().putString(key, str).apply()
+    }
+
+    fun isKeyExist(): Boolean = accessPreference.getString("Token", "").let {
+        !it.isNullOrEmpty()
+    }
+
+    fun deleteKey() {
+        setString(str = "")
+    }
+}
