+++ data/src/main/java/com/gyleedev/data/remote/Qualifiers.kt
@@ -1,0 +1,15 @@
+package com.gyleedev.data.remote
+
+import javax.inject.Qualifier
+
+@Qualifier
+@Retention(AnnotationRetention.BINARY)
+annotation class TypeAccess
+
+@Qualifier
+@Retention(AnnotationRetention.BINARY)
+annotation class TypeApi
+
+@Qualifier
+@Retention(AnnotationRetention.BINARY)
+annotation class TypeRevoke
