+++ data/src/main/java/com/gyleedev/data/remote/response/GithubAccessResponse.kt
@@ -1,0 +1,5 @@
+package com.gyleedev.data.remote.response
+
+import com.google.gson.annotations.SerializedName
+
+data class GithubAccessResponse(@SerializedName("access_token") val accessToken: String)
