+++ data/src/main/java/com/gyleedev/data/remote/response/RevokeResponse.kt
@@ -1,0 +1,5 @@
+package com.gyleedev.data.remote.response
+
+import com.google.gson.annotations.SerializedName
+
+data class RevokeResponse(@SerializedName("Status") val status: Int)
