+++ data/src/main/java/com/gyleedev/data/repository/RepositoryModule.kt
@@ -1,0 +1,17 @@
+package com.gyleedev.data.repository
+
+import com.gyleedev.githubsearch.domain.repository.GitHubRepository
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+abstract class RepositoryModule {
+
+    @Binds
+    @Singleton
+    abstract fun bindsUserRepository(impl: GitHubRepositoryImpl): GitHubRepository
+}
